% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwitterLanguageVisualizer.R
\name{my_function}
\alias{my_function}
\title{Twitter Language Visualizer}
\usage{
my_function(search = "cats", num_tweets = 100, color = "#aecca5")
}
\arguments{
\item{search}{A term that the user would like to search on Twitter.}

\item{num_tweets}{The number of tweets the user would like to search on
Twitter.}

\item{color}{A hexidecimal color that the user would like the bar graph
visualization to be.}
}
\value{
The Twitter language count in bar graph form from user chosen search
values or default values provided.
}
\description{
Visualizes Languages Used On Twitter
}
\details{
A function that visualizes the number of languages used on Twitter in a bar
graph based off of user chosen terms for search value, number of tweets and
hexidecimal bar graph color.
}
\examples{
my_function(search = "pizza", num_tweets = 125, color = "#c3a7d6")
my_function(search = "hawaii", num_tweets = 250, color = "#f09cc6")

}
\author{
McKinley O'Connor
}
